Finished Some Enemy AI
