+++ Assets/Scenes/Enemies/Components.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 86910860a6bbdd5498bcf58c38a98e1c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
