+++ Assets/Scenes/Enemies/Components/WallDetection.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 913aa9d5f3fa7104a8c25a12ca442878
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
