+++ Assets/Scenes/Enemies/Enemy1.prefab
@@ -1,35 +1,5 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &6146904609052278757
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 6146904609052278756}
-  m_Layer: 0
-  m_Name: GroundDetection
-  m_TagString: Component
-  m_Icon: {fileID: 2488908585195742037, guid: 0000000000000000d000000000000000, type: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &6146904609052278756
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6146904609052278757}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0.103, y: -0.035, z: -1.6809533}
-  m_LocalScale: {x: 0.25, y: 0.25, z: 1}
-  m_Children: []
-  m_Father: {fileID: 2253572787408656668}
-  m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &6146904609136861509
 GameObject:
   m_ObjectHideFlags: 0
@@ -40,9 +10,9 @@ GameObject:
   m_Component:
   - component: {fileID: 2253572787408656668}
   - component: {fileID: 6146904609136861369}
-  - component: {fileID: 6146904609136861511}
   - component: {fileID: 6146904609136861510}
   - component: {fileID: 6146904609136861508}
+  - component: {fileID: 6146904609136861511}
   m_Layer: 0
   m_Name: Enemy1
   m_TagString: Enemy
@@ -115,21 +85,6 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!114 &6146904609136861511
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6146904609136861509}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 98aab6edc593e8149937dfc6eb79e04f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 3
-  distance: 5
-  groundDetection: {fileID: 6146904609052278756}
 --- !u!50 &6146904609136861510
 Rigidbody2D:
   serializedVersion: 4
@@ -177,3 +132,93 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 0.2, y: 0.2}
   m_EdgeRadius: 0
+--- !u!114 &6146904609136861511
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6146904609136861509}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 98aab6edc593e8149937dfc6eb79e04f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  speed: 3
+  distance: 0.1
+  wallDetection: {fileID: 6146904609052278756}
+--- !u!1001 &6612838730519872664
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 2253572787408656668}
+    m_Modifications:
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0.103
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -0.035
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -1.6809533
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1048130240835009405, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+        type: 3}
+      propertyPath: m_Name
+      value: WallDetection
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 913aa9d5f3fa7104a8c25a12ca442878, type: 3}
+--- !u!4 &6146904609052278756 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 1048130240835009404, guid: 913aa9d5f3fa7104a8c25a12ca442878,
+    type: 3}
+  m_PrefabInstance: {fileID: 6612838730519872664}
+  m_PrefabAsset: {fileID: 0}
