--- Assets/Scripts/Enemy1Movement.cs
@@ -1,57 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Enemy1Movement : MonoBehaviour
-{
-    public float speed;
-    public float distance;
-
-    private bool movingRight = true;
-
-    public Transform groundDetection;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        Physics2D.gravity = new Vector2(0, -9.8f);
-
-        transform.Translate(Vector2.right * speed * Time.deltaTime);
-
-        RaycastHit2D groundInfo = Physics2D.Raycast(groundDetection.position, Vector2.down, distance);
-        if (groundInfo.collider == false)
-        {
-            if (movingRight == true)
-            {
-                transform.eulerAngles = new Vector3(0, -180, 0);
-                movingRight = false;
-            }
-            else
-            {
-                transform.eulerAngles = new Vector3(0, 0, 0);
-                movingRight = true;
-            }
-        }
-    }
-
-    private void OnTriggerEnter2D(Collider2D other)
-    {
-        // The enemy keeps getting stuck on a wall
-        if (movingRight == true)
-        {
-            transform.eulerAngles = new Vector3(0, -180, 0);
-            movingRight = false;
-        }
-        else
-        {
-            transform.eulerAngles = new Vector3(0, 0, 0);
-            movingRight = true;
-        }
-    }
-}
