+++ Assets/Scripts/Enemy2Script.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46bda1f1c8cd1d94396e52ed3090715d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
