+++ Assets/Scripts/SpikeScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5526fa3fdf031514aa08f75972cfa78f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
